Created initial file upload functionality with backend. Backend can only maintain 1 demo for now
