+++ package.json
@@ -7,6 +7,7 @@
     "typescript": "^5.8.3"
   },
   "dependencies": {
+    "react-router-dom": "^7.5.1",
     "react-svg": "^16.3.0"
   }
 }
